Optimize memory usage for embedded systems
- Implement dynamic allocation for per-channel structures (CoderInfo,
  ChannelInfo, PsyInfo) based on the actual number of channels requested
  at runtime, reducing overhead for mono/stereo streams.
- Change quantized spectrum storage from 32-bit int to 16-bit int16_t,
  halving memory usage for spectral buffers while maintaining standard
  compliance (coeffs capped at 8191).
- Update SSE2 quantization kernel with _mm_packs_epi32 for safe 16-bit
  packing with saturation.
- Reduce TnsInfo memory footprint by decreasing windowData array size
  from 8 to 1, reflecting that TNS is currently only used for long blocks.
- Update TNS and bitstream writing logic to handle the reduced window
  data array safely.
- Verified with mono, stereo, and multichannel encoding tests.

+++ libfaac/huff2.h
@@ -64,7 +64,7 @@ static inline int clamp_sf_diff(int diff)
 }
 
 int huffbook(CoderInfo *coderInfo,
-             int *qs /* quantized spectrum */,
+             int16_t *qs /* quantized spectrum */,
              int len);
 int writebooks(CoderInfo *coder, BitStream *stream, int writeFlag);
 int writesf(CoderInfo *coder, BitStream *bitStream, int writeFlag);
